tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/049.html

<!DOCTYPE html>
<html><head>
  <title> scheduler: adding external script but removeAttribute( src ) before it runs</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<body>

  <div id="log">FAILED (This TC requires JavaScript enabled)</div>

  <script>log('inline script #1');
    var script=testlib.addScript('log(\'HEAD script\');', { src:'scripts/include-1.js' }, document.getElementsByTagName('head')[0], false );
    script.removeAttribute('src');
    log('end script #1');
  </script>

  <script type="text/javascript">
  var t = async_test()

  function test() {
    assert_array_equals(eventOrder, ['inline script #1', 'end script #1', 'external script #1']);
    t.done();
  }
  onload = t.step_func(test)
  </script>

</body></html>
